Kraken API changes
Due to changes to the now deprecated Twitch Kraken API, users have
reported receiving an error message when attempting to use the CLI (#7);
This is because as of the following reasons:

1. Twitch changed the way streams were requested, requiring a user ID
instead of a login name.
2. As of May 31, 2019, Twitch changed the "Accept" header required to
interact with the legacy API.

This has been changed by first requesting the user's ID, and if
successful, then requesting the stream; This resolves the error and
allows the CLI to function as expected.